+++ themegen/lib/src/feature/ext/logic/ext_producer.dart
@@ -1,12 +1,14 @@
 import 'package:analyzer/dart/element/element.dart';
-import 'package:analyzer/dart/element/type.dart';
 import 'package:code_builder/code_builder.dart';
 import 'package:themegen/src/core/logic/code_producer.dart';
-import 'package:themegen/src/core/utils/extension.dart';
 import 'package:themegen/src/feature/ext/logic/convert.dart';
-import 'package:themegen/src/feature/ext/logic/extension.dart';
+import 'package:themegen/src/feature/ext/logic/extension_copywith.dart';
+import 'package:themegen/src/feature/ext/logic/extension_lerp.dart';
+import 'package:themegen/src/feature/ext/logic/extension_method.dart';
+import 'package:themegen/src/feature/ext/logic/lerp_constructor.dart';
+import 'package:themegen/src/feature/ext/logic/params.dart';
 import 'package:themegen/src/feature/ext/model/elements.dart';
-import 'package:themegen/src/feature/ext/model/extension.dart';
+import 'package:themegen/src/feature/ext/model/extension_model.dart';
 
 class ExtProducer extends CodeProducer<ParentExtension> {
   ExtProducer(super.emitter);
@@ -55,18 +57,19 @@ class ExtProducer extends CodeProducer<ParentExtension> {
           _generateFields($allFields),
         )
         ..methods.addAll([
+          /// generate all extension methods, i.e. light, dark, etc.
           ...extensions.map(
             (typeExtension) => extensionMethod(
               className,
               typeExtension,
             ),
           ),
-          _lerp(
+          lerp(
             className,
             $onlyFields,
             $onlyMethods,
           ),
-          _copyWith(
+          extensionCopywith(
             className,
             $onlyFields,
             $onlyMethods,
@@ -77,7 +80,7 @@ class ExtProducer extends CodeProducer<ParentExtension> {
             Constructor(
               (builder) => builder
                 ..optionalParameters.addAll(
-                  _params(
+                  paramsFromFields(
                     $onlyFields,
                     type,
                     required: true,
@@ -85,13 +88,13 @@ class ExtProducer extends CodeProducer<ParentExtension> {
                   ),
                 )
                 ..optionalParameters.addAll(
-                  _paramsFromMethods($onlyMethods),
+                  paramsFromMethods($onlyMethods),
                 )
                 ..initializers.addAll(
                   _initializersFromMethods($onlyMethods),
                 ),
             ),
-            _lerpConstructor(
+            lerpConstructor(
               $onlyFields,
               $onlyMethods,
               type,
@@ -100,63 +103,6 @@ class ExtProducer extends CodeProducer<ParentExtension> {
           ],
         ),
     );
-  }
-
-  Constructor _lerpConstructor(
-    Iterable<ExtensionField> fields,
-    Iterable<ExtensionMethod> methods,
-    DartType type,
-    String className,
-  ) =>
-      Constructor(
-        (builder) => builder
-          ..name = '_lerp'
-          ..optionalParameters.addAll([
-            ..._params(
-              fields,
-              type,
-              required: true,
-            ),
-            ..._paramsFromMethods(methods),
-            Parameter(
-              (builder) => builder
-                ..name = 't'
-                ..required = true
-                ..type = refer('double'),
-            ),
-            Parameter(
-              (builder) => builder
-                ..name = 'other'
-                ..required = true
-                ..type = refer(className),
-            ),
-          ])
-          ..initializers.addAll(
-            _lerpConstructorInitializers(
-              fields,
-              methods,
-            ),
-          ),
-      );
-
-  Iterable<Code> _lerpConstructorInitializers(
-    Iterable<ExtensionField> fields,
-    Iterable<ExtensionMethod> methods,
-  ) sync* {
-    for (final field in fields) {
-      final name = field.name;
-      yield Code('$name = ${field.type}.lerp($name, other.$name, t)!');
-    }
-    for (final method in methods) {
-      final name = method.name;
-      final params = method.parameters;
-      final onlyNames = params.map((p) => p.name).join(', ');
-      yield Code(
-        '$name = ${method.returnType}.lerp($name($onlyNames), other._\$$name($onlyNames), t)!',
-      );
-      yield Code('_\$$name = $name');
-      yield* params.toSet().map((param) => Code('${param.name} = ${param.name}'));
-    }
   }
 
   Iterable<Code> _initializersFromMethods(
@@ -177,37 +123,6 @@ class ExtProducer extends CodeProducer<ParentExtension> {
         );
   }
 
-  Iterable<Parameter> _paramsFromMethods(
-    Iterable<ExtensionMethod> methods, {
-    bool required = true,
-  }) sync* {
-    yield* methods.map(
-      (m) {
-        final returnType = m.returnType;
-        final params = m.parameters;
-        final types = params.map((e) => e.type).join(', ');
-        return Parameter(
-          (builder) => builder
-            ..name = m.name
-            ..named = true
-            ..required = required
-            ..type = refer(
-              '$returnType Function($types)${required ? '' : '?'}',
-            ),
-        );
-      },
-    );
-    yield* methods.expand((element) => element.parameters).map(
-          (e) => Parameter(
-            (builder) => builder
-              ..name = e.name
-              ..named = true
-              ..type = refer(e.type)
-              ..required = true,
-          ),
-        );
-  }
-
   Iterable<Field> _generateFields(
     Iterable<ExtensionField> fields,
   ) =>
@@ -218,119 +133,5 @@ class ExtProducer extends CodeProducer<ParentExtension> {
             ..name = field.name
             ..type = refer(field.type),
         ),
-      );
-
-  Iterable<Parameter> _params(
-    Iterable<ExtensionField> fields,
-    DartType type, {
-    bool named = true,
-    bool required = false,
-    bool toThis = false,
-    bool defaultTo = false,
-  }) =>
-      fields.map(
-        (field) => Parameter(
-          (builder) => builder
-            ..name = field.name
-            ..named = named
-            ..required = required
-            ..toThis = toThis
-            ..defaultTo = defaultTo
-                ? Code(
-                    type.getDisplayString(withNullability: false) + '.${field.name}',
-                  )
-                : null
-            ..type = toThis
-                ? null
-                : refer(
-                    '${field.type}${required ? '' : '?'}',
-                  ),
-        ),
-      );
-
-  Method _lerp(
-    String className,
-    Iterable<ExtensionField> fields,
-    Iterable<ExtensionMethod> methods,
-  ) =>
-      Method(
-        (builder) => builder
-          ..name = 'lerp'
-          ..annotations.add(refer('override'))
-          ..returns = refer(className)
-          ..requiredParameters.addAll([
-            Parameter(
-              (builder) => builder
-                ..name = 'other'
-                ..type = refer('ThemeExtension<$className>?'),
-            ),
-            Parameter(
-              (builder) => builder
-                ..name = 't'
-                ..type = refer('double'),
-            ),
-          ])
-          ..body = Code('''
-            if (other is! $className) {
-              return this;
-            }
-            return $className._lerp(
-              other: other,
-              t: t,
-              ${fields.map((e) => '${e.name}: ${e.name}').joinParams(',')}
-              ${methods.map((e) => '${e.name}: _\$${e.name}').joinParams(',')}
-              ${methods.expand((e) => e.parameters).map((e) => '${e.name}: ${e.name}').joinParams(',')}
-            );
-            '''),
-      );
-
-  Method _copyWith(
-    String className,
-    Iterable<ExtensionField> fields,
-    Iterable<ExtensionMethod> methods,
-  ) =>
-      Method(
-        (builder) => builder
-          ..name = 'copyWith'
-          ..returns = refer(className)
-          ..annotations.add(refer('override'))
-          ..body = Code('''
-          return $className(
-            ${fields.map<String>((f) => '${f.name}: ${f.name} ?? this.${f.name}').join(',')}
-            ${methods.map((m) => '${m.name}: ${m.name} ?? _\$${m.name}').join(',')},
-            ${methods.expand((m) => m.parameters).map((p) => '${p.name}: ${p.name} ?? this.${p.name}').join(',')}
-          );
-        ''')
-          ..optionalParameters.addAll([
-            for (final field in fields)
-              Parameter(
-                (builder) => builder
-                  ..name = field.name
-                  ..named = true
-                  ..type = refer(
-                    '${field.type}?',
-                  ),
-              ),
-            for (final method in methods)
-              Parameter(
-                (builder) {
-                  final params = method.parameters.map((method) => method.type).join(', ');
-                  builder
-                    ..name = method.name
-                    ..named = true
-                    ..type = refer(
-                      '${method.returnType} Function($params)?',
-                    );
-                },
-              ),
-            ...methods.expand((method) => method.parameters).map(
-                  (param) => Parameter(
-                    (builder) => builder
-                      ..name = param.name
-                      ..named = true
-                      ..type = refer('${param.type}?'),
-                  ),
-                )
-          ]),
       );
 }
